Changed datatype of price from int to long for better coorellation with firebase's number datatype.

+++ app/src/main/java/com/masthuggis/boki/presenter/CreateAdPresenter.java
@@ -115,7 +115,7 @@ public class CreateAdPresenter{
         return advertisement.getDescription();
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return advertisement.getPrice();
     }
